Sim2real environment constructed and Sim2sim testing completed in the Gazebo environment.
