+++ scripts/sim2real/OMY/reach/utils/policy_controller.py
@@ -1,33 +1,32 @@
 #!/usr/bin/env python3
-"""
-policy_controller.py
---------------------
 
-Minimal wrapper to load a TorchScript policy and expose helpers to:
+# Copyright 2025 ROBOTIS CO., LTD.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# Author: Taehyeong Kim
 
-* build observations  – implemented in subclasses
-* compute actions     – via `_compute_action()`
-
-The class also extracts physics and robot-joint parameters from the
-`env.yaml` that accompanies each policy.
-
-Sub-classes must set `self.dof_names` before calling `load_policy`
-(in `__init__`) and implement:
-
-* `_compute_observation()`
-* `forward()`
-
-Author: Louis Le Lay
-"""
+# Modified from original code by Louis Le Lay
+# Original source: https://github.com/louislelay/kinova_isaaclab_sim2real
 
 import io
 
 import numpy as np
 import torch
 
-from utils.config_loader import parse_env_config, get_physics_properties, get_robot_joint_properties
+from .config_loader import parse_env_config, get_physics_properties, get_robot_joint_properties
 
-class PolicyController:
+class PolicyExecutor:
     """A controller that loads and executes a policy from a file."""
 
     def __init__(self) -> None:
